test(core): refactor ut
