+++ packages/fx-core/src/core/environment.ts
@@ -23,6 +23,7 @@ import {
 import path, { basename } from "path";
 import fs from "fs-extra";
 import jsum from "jsum";
+import * as dotenv from "dotenv";
 import {
   deserializeDict,
   dataNeedEncryption,
@@ -314,7 +315,7 @@ class EnvironmentManager {
       : this.getConfigFolder(projectPath);
   }
 
-  private getEnvConfigsFolder(projectPath: string): string {
+  public getEnvConfigsFolder(projectPath: string): string {
     return path.resolve(this.getConfigFolder(projectPath), InputConfigsFolderName);
   }
