+++ packages/fx-core/src/core/middleware/concurrentLocker.ts
@@ -4,6 +4,7 @@
 
 import { HookContext, Middleware, NextFunction } from "@feathersjs/hooks";
 import {
+  ConcurrentError,
   ConfigFolderName,
   CoreCallbackEvent,
   err,
@@ -12,19 +13,14 @@ import {
   StaticPlatforms,
 } from "@microsoft/teamsfx-api";
 import * as fs from "fs-extra";
-import * as os from "os";
+
 import * as path from "path";
 import { lock, unlock } from "proper-lockfile";
 import { FxCore } from "..";
 import { CallbackRegistry } from "../callback";
-import {
-  ConcurrentError,
-  InvalidProjectError,
-  NoProjectOpenedError,
-  PathNotExistError,
-} from "../error";
-import { base64Encode } from "../tools";
-
+import { CoreSource, InvalidProjectError, NoProjectOpenedError, PathNotExistError } from "../error";
+import { getLockFolder } from "../tools";
+import crypto from "crypto";
 export const ConcurrentLockerMW: Middleware = async (ctx: HookContext, next: NextFunction) => {
   const core = ctx.self as FxCore;
   const inputs = ctx.arguments[ctx.arguments.length - 1] as Inputs;
@@ -42,40 +38,39 @@ export const ConcurrentLockerMW: Middleware = async (ctx: HookContext, next: Nex
       ctx.result = err(PathNotExistError(inputs.projectPath));
       return;
     }
-    const lf = path.join(inputs.projectPath, `.${ConfigFolderName}`);
-    if (!(await fs.pathExists(lf))) {
+    const configFolder = path.join(inputs.projectPath, `.${ConfigFolderName}`);
+    if (!(await fs.pathExists(configFolder))) {
       ctx.result = err(InvalidProjectError());
       return;
     }
 
-    const lockFileDir = path.join(
-      os.tmpdir(),
-      `${ProductName}-${base64Encode(inputs.projectPath)}`
-    );
+    const lockFileDir = getLockFolder(inputs.projectPath);
     await fs.ensureDir(lockFileDir);
 
-    await lock(lf, { lockfilePath: path.join(lockFileDir, `${ConfigFolderName}.lock`) })
+    await lock(configFolder, { lockfilePath: path.join(lockFileDir, `${ConfigFolderName}.lock`) })
       .then(async () => {
-        if (logger) logger.debug(`[core] success to acquire lock on: ${lf}`);
+        if (logger) logger.debug(`[core] success to acquire lock on: ${configFolder}`);
         for (const f of CallbackRegistry.get(CoreCallbackEvent.lock)) {
           f();
         }
         try {
           await next();
         } finally {
-          await unlock(lf, { lockfilePath: path.join(lockFileDir, `${ConfigFolderName}.lock`) });
-          await fs.rmdir(lockFileDir);
+          await unlock(configFolder, {
+            lockfilePath: path.join(lockFileDir, `${ConfigFolderName}.lock`),
+          });
+          await fs.rmdir(lockFileDir, { recursive: true });
 
           for (const f of CallbackRegistry.get(CoreCallbackEvent.unlock)) {
             f();
           }
-          if (logger) logger.debug(`[core] lock released on ${lf}`);
+          if (logger) logger.debug(`[core] lock released on ${configFolder}`);
         }
       })
       .catch((e: any) => {
         if (e["code"] === "ELOCKED") {
-          if (logger) logger.warning(`[core] failed to acquire lock on: ${lf}`);
-          ctx.result = err(new ConcurrentError());
+          if (logger) logger.warning(`[core] failed to acquire lock on: ${configFolder}`);
+          ctx.result = err(new ConcurrentError(CoreSource));
           return;
         }
         throw e;
