+++ packages/fx-core/src/core/middleware/index.ts
@@ -1,0 +1,15 @@
+export * from "./concurrentLocker";
+export * from "./contextInjector";
+export * from "./envInfoLoader";
+export * from "./envInfoWriter";
+export * from "./errorHandler";
+export * from "./localSettingsLoader";
+export * from "./localSettingsWriter";
+export * from "./migrateConditionHandler";
+export * from "./projectMigrator";
+export * from "./projectSettingsLoader";
+export * from "./projectSettingsWriter";
+export * from "./projectUpgrader";
+export * from "./questionModel";
+export * from "./solutionLoader";
+export * from "./telemetrySender";
