+++ packages/fx-core/src/core/middleware/projectSettingsWriter.ts
@@ -55,8 +55,8 @@ export const ProjectSettingsWriterMW: Middleware = async (
     await fs.writeFile(settingFile, JSON.stringify(projectSettings, null, 4));
     core.tools.logProvider.debug(`[core] persist project setting file: ${settingFile}`);
   } catch (e) {
-    if ((ctx.res as Result<any, FxError>).isOk()) {
-      ctx.res = err(WriteFileError(e));
+    if ((ctx.result as Result<any, FxError>).isOk()) {
+      ctx.result = err(WriteFileError(e));
     }
   }
 };
