+++ packages/fx-core/tests/core/hooks.test.ts
@@ -1,44 +1,30 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { hooks, Middleware, NextFunction } from "@feathersjs/hooks/lib";
+import { hooks, NextFunction } from "@feathersjs/hooks/lib";
 import {
   AppPackageFolderName,
   ArchiveFolderName,
-  AzureSolutionSettings,
   ConfigFolderName,
-  ConfigMap,
-  CoreCallbackEvent,
   EnvConfig,
   EnvInfo,
   Err,
   err,
-  Func,
-  FunctionRouter,
   FxError,
-  InputConfigsFolderName,
   Inputs,
-  InputTextConfig,
   Json,
   Ok,
   ok,
   Platform,
-  ProductName,
-  ProjectSettingsFileName,
-  QTreeNode,
   Result,
   SingleSelectConfig,
   SingleSelectResult,
-  Solution,
   SolutionContext,
   Stage,
-  SystemError,
   Tools,
   UserCancelError,
-  UserError,
 } from "@microsoft/teamsfx-api";
-import { assert, expect } from "chai";
-import * as dotenv from "dotenv";
+import { assert } from "chai";
 import fs from "fs-extra";
 import "mocha";
 import mockedEnv from "mocked-env";
@@ -46,46 +32,19 @@ import * as os from "os";
 import * as path from "path";
 import sinon from "sinon";
 import { Container } from "typedi";
-import * as uuid from "uuid";
-import {
-  base64Encode,
-  CoreHookContext,
-  deserializeDict,
-  InvalidInputError,
-  isV2,
-  mapToJson,
-  serializeDict,
-} from "../../src";
+import { CoreHookContext, deserializeDict, serializeDict } from "../../src";
 import { FeatureFlagName } from "../../src/common/constants";
 import * as commonTools from "../../src/common/tools";
-import { CallbackRegistry } from "../../src/core/callback";
 import { environmentManager } from "../../src/core/environment";
-import {
-  ConcurrentError,
-  InvalidProjectError,
-  NoProjectOpenedError,
-  PathNotExistError,
-} from "../../src/core/error";
-import { ConcurrentLockerMW } from "../../src/core/middleware/concurrentLocker";
+import { NoProjectOpenedError } from "../../src/core/error";
 import { ContextInjectorMW } from "../../src/core/middleware/contextInjector";
 import { EnvInfoLoaderMW } from "../../src/core/middleware/envInfoLoader";
-import { EnvInfoWriterMW } from "../../src/core/middleware/envInfoWriter";
-import { ErrorHandlerMW } from "../../src/core/middleware/errorHandler";
 import { LocalSettingsLoaderMW } from "../../src/core/middleware/localSettingsLoader";
 import { MigrateConditionHandlerMW } from "../../src/core/middleware/migrateConditionHandler";
 import { migrateArm, ProjectMigratorMW } from "../../src/core/middleware/projectMigrator";
-import {
-  newSolutionContext,
-  ProjectSettingsLoaderMW,
-} from "../../src/core/middleware/projectSettingsLoader";
-import { ProjectSettingsWriterMW } from "../../src/core/middleware/projectSettingsWriter";
 import { ProjectUpgraderMW } from "../../src/core/middleware/projectUpgrader";
-import { QuestionModelMW } from "../../src/core/middleware/questionModel";
-import { SolutionLoaderMW } from "../../src/core/middleware/solutionLoader";
 import { TelemetrySenderMW } from "../../src/core/middleware/telemetrySender";
 import { SolutionPlugins } from "../../src/core/SolutionPluginContainer";
-import { PluginNames } from "../../src/plugins/solution/fx-solution/constants";
-import { AzureResourceSQL } from "../../src/plugins/solution/fx-solution/question";
 import {
   MockLatestVersion2_3_0Context,
   MockLatestVersion2_3_0UserData,
@@ -98,15 +57,8 @@ import {
   randomAppName,
 } from "./utils";
 
-const EnvParams = [
-  { TEAMSFX_APIV2: "false", TEAMSFX_MULTI_ENV: "false" },
-  { TEAMSFX_APIV2: "false", TEAMSFX_MULTI_ENV: "true" },
-  { TEAMSFX_APIV2: "true", TEAMSFX_MULTI_ENV: "false" },
-  { TEAMSFX_APIV2: "true", TEAMSFX_MULTI_ENV: "true" },
-];
-
 let mockedEnvRestore: () => void;
-describe("Middleware", () => {
+describe("Middleware - others", () => {
   const sandbox = sinon.createSandbox();
   const mockSolution = new MockSolution();
 
@@ -115,1042 +67,6 @@ describe("Middleware", () => {
   });
   afterEach(() => {
     sandbox.restore();
-  });
-  describe("ErrorHandlerMW", () => {
-    const inputs: Inputs = { platform: Platform.VSCode };
-
-    it("return error", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return err(UserCancelError);
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ErrorHandlerMW],
-      });
-      const my = new MyClass();
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isErr() && res.error === UserCancelError);
-    });
-
-    it("return ok", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("hello");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ErrorHandlerMW],
-      });
-      const my = new MyClass();
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isOk() && res.value === "hello");
-      my.tools = undefined;
-      const res2 = await my.myMethod(inputs);
-      assert.isTrue(res2.isOk() && res2.value === "hello");
-    });
-
-    it("throw known error", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          throw UserCancelError;
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ErrorHandlerMW],
-      });
-      const my = new MyClass();
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isErr() && res.error === UserCancelError);
-    });
-
-    it("throw unknown error", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          throw new Error("unknown");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ErrorHandlerMW],
-      });
-      const my = new MyClass();
-      const res = await my.myMethod(inputs);
-      assert.isTrue(
-        res.isErr() && res.error instanceof SystemError && res.error.message === "unknown"
-      );
-    });
-
-    it("convert system error to user error", async () => {
-      const msg =
-        "The client 'xxx@xxx.com' with object id 'xxx' does not have authorization to perform action '<REDACTED: user-file-path>' over scope '<REDACTED: user-file-path>' or the scope is invalid. If access was recently granted, please refresh your credentials.";
-
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          throw new Error(msg);
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ErrorHandlerMW],
-      });
-      const my = new MyClass();
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isErr());
-      if (res.isErr()) {
-        const error = res.error;
-        assert.isTrue(error instanceof UserError);
-        assert.equal(error.message, msg);
-      }
-    });
-  });
-  describe("ConcurrentLockerMW", () => {
-    it("temp folder should be existed when it's locked", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          const lockFileDir = path.join(
-            os.tmpdir(),
-            `${ProductName}-${base64Encode(inputs.projectPath!)}`
-          );
-          expect(await fs.pathExists(lockFileDir)).is.true;
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        const res = await my.myMethod(inputs);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("temp folder should be removed after being unlocked", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        const res = await my.myMethod(inputs);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-      const lockFileDir = path.join(
-        os.tmpdir(),
-        `${ProductName}-${base64Encode(inputs.projectPath!)}`
-      );
-      expect(await fs.pathExists(lockFileDir)).is.false;
-    });
-
-    it("sequence: ok", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        const res = await my.myMethod(inputs);
-        assert.isTrue(res.isOk() && res.value === "");
-        my.tools = undefined;
-        const res2 = await my.myMethod(inputs);
-        assert.isTrue(res2.isOk() && res2.value === "");
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("single: throw error", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          throw UserCancelError;
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        await my.myMethod(inputs);
-      } catch (e) {
-        assert.isTrue(e === UserCancelError);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("single: invalid NoProjectOpenedError", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = undefined;
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isErr() && res.error.name === NoProjectOpenedError().name);
-    });
-
-    it("single: invalid PathNotExistError", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isErr() && res.error.name === PathNotExistError(inputs.projectPath).name);
-    });
-
-    it("single: invalid InvalidProjectError", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        const res = await my.myMethod(inputs);
-        assert.isTrue(res.isErr() && res.error.name === InvalidProjectError().name);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("concurrent: fail to get lock", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          const res = await this.myMethod(inputs);
-          assert.isTrue(res.isErr() && res.error.name === new ConcurrentError().name);
-          this.tools = undefined;
-          const res2 = await this.myMethod(inputs);
-          assert.isTrue(res2.isErr() && res2.error.name === new ConcurrentError().name);
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const my = new MyClass();
-      try {
-        inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        await my.myMethod(inputs);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("concurrent: ignore lock", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          const inputs2: Inputs = { platform: Platform.VSCode, ignoreLock: true };
-          const res2 = await this.myMethod2(inputs2);
-          assert.isTrue(res2.isOk() && res2.value === "");
-          return ok("");
-        }
-
-        async myMethod2(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-        myMethod2: [ConcurrentLockerMW],
-      });
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const my = new MyClass();
-      try {
-        inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        await my.myMethod(inputs);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-
-    it("callback should work", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      let d = 0;
-
-      const lockCb = () => {
-        d++;
-      };
-
-      const unlockCb = () => {
-        d--;
-      };
-
-      CallbackRegistry.set(CoreCallbackEvent.lock, lockCb);
-      CallbackRegistry.set(CoreCallbackEvent.lock, lockCb);
-      CallbackRegistry.set(CoreCallbackEvent.unlock, unlockCb);
-
-      hooks(MyClass, {
-        myMethod: [ConcurrentLockerMW],
-      });
-
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      try {
-        await fs.ensureDir(inputs.projectPath);
-        await fs.ensureDir(path.join(inputs.projectPath, `.${ConfigFolderName}`));
-        await my.myMethod(inputs);
-        expect(d).eql(1);
-      } finally {
-        await fs.rmdir(inputs.projectPath!, { recursive: true });
-      }
-    });
-  });
-
-  describe("SolutionLoaderMW, ContextInjectorMW", () => {
-    const MockProjectSettingsMW = async (ctx: CoreHookContext, next: NextFunction) => {
-      ctx.projectSettings = {
-        appName: "testapp",
-        projectId: uuid.v4(),
-        version: "2.0.0",
-        solutionSettings: {
-          name: "fx-solution-azure",
-        },
-      };
-      await next();
-    };
-
-    it("load solution from zero and inject (API V1)", async () => {
-      const mockedEnvRestore = mockedEnv({ TEAMSFX_APIV2: "false" });
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.solution !== undefined);
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [SolutionLoaderMW(), ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isOk() && res.value === "");
-      mockedEnvRestore();
-    });
-    it("load solution from zero and inject (API V2)", async () => {
-      const mockedEnvRestore = mockedEnv({ TEAMSFX_APIV2: "true" });
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.solutionV2 !== undefined);
-          return ok("");
-        }
-      }
-      hooks(MyClass, {
-        myMethod: [SolutionLoaderMW(), ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isOk() && res.value === "");
-      mockedEnvRestore();
-    });
-    it("load solution from existing project and inject (API V1)", async () => {
-      const mockedEnvRestore = mockedEnv({ TEAMSFX_APIV2: "false" });
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.solution !== undefined);
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [MockProjectSettingsMW, SolutionLoaderMW(), ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isOk() && res.value === "");
-      mockedEnvRestore();
-    });
-    it("load solution from existing project and inject (API V2)", async () => {
-      const mockedEnvRestore = mockedEnv({ TEAMSFX_APIV2: "true" });
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.solutionV2 !== undefined);
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [MockProjectSettingsMW, SolutionLoaderMW(), ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const res = await my.myMethod(inputs);
-      assert.isTrue(res.isOk() && res.value === "");
-      mockedEnvRestore();
-    });
-  });
-
-  describe("ProjectSettingsLoaderMW, ContextInjectorMW: part 1", () => {
-    it("ignore loading project settings", async () => {
-      class MyClass {
-        tools = new MockTools();
-        async getQuestions(
-          stage: Stage,
-          inputs: Inputs,
-          ctx?: CoreHookContext
-        ): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.projectSettings === undefined);
-          return ok("");
-        }
-        async other(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined && ctx.projectSettings === undefined);
-          return ok("");
-        }
-      }
-      hooks(MyClass, {
-        getQuestions: [ProjectSettingsLoaderMW, ContextInjectorMW],
-        other: [ProjectSettingsLoaderMW, ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      await my.getQuestions(Stage.create, inputs);
-      inputs.platform = Platform.CLI_HELP;
-      await my.other(inputs);
-      inputs.platform = Platform.VS;
-      await my.other(inputs);
-    });
-
-    it("failed to load: NoProjectOpenedError, PathNotExistError", async () => {
-      class MyClass {
-        tools?: any = new MockTools();
-        async other(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-      hooks(MyClass, {
-        other: [ProjectSettingsLoaderMW, ContextInjectorMW],
-      });
-      const my = new MyClass();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const res = await my.other(inputs);
-      assert.isTrue(res.isErr() && res.error.name === NoProjectOpenedError().name);
-      inputs.projectPath = path.join(os.tmpdir(), randomAppName());
-      const res2 = await my.other(inputs);
-      assert.isTrue(res2.isErr() && res2.error.name === PathNotExistError(inputs.projectPath).name);
-    });
-  });
-
-  describe("ProjectSettingsLoaderMW, ContextInjectorMW: part 2", () => {
-    const sandbox = sinon.createSandbox();
-    const appName = randomAppName();
-    const projectSettings = MockProjectSettings(appName);
-    const inputs: Inputs = { platform: Platform.VSCode };
-    inputs.projectPath = path.join(os.tmpdir(), appName);
-    const confFolderPath = path.resolve(inputs.projectPath, `.${ConfigFolderName}`);
-    const settingsFiles = [
-      path.resolve(confFolderPath, "settings.json"),
-      path.resolve(confFolderPath, InputConfigsFolderName, ProjectSettingsFileName),
-    ];
-
-    beforeEach(() => {
-      sandbox.stub<any, any>(fs, "readJson").callsFake(async (file: string) => {
-        if (settingsFiles.includes(file)) return projectSettings;
-        return undefined;
-      });
-      sandbox.stub<any, any>(fs, "pathExists").callsFake(async (file: string) => {
-        if (settingsFiles.includes(file)) return true;
-        if (inputs.projectPath === file) return true;
-        return false;
-      });
-    });
-
-    afterEach(() => {
-      sandbox.restore();
-    });
-
-    for (const param of EnvParams) {
-      describe(`Multi-Env: ${param.TEAMSFX_MULTI_ENV}, API V2:${param.TEAMSFX_APIV2}`, () => {
-        beforeEach(() => {
-          mockedEnvRestore = mockedEnv(param);
-        });
-
-        afterEach(() => {
-          mockedEnvRestore();
-        });
-
-        it(`success to load project settings`, async () => {
-          class MyClass {
-            tools = new MockTools();
-            async other(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-              assert.isTrue(ctx !== undefined);
-              if (ctx) {
-                assert.deepEqual(projectSettings, ctx.projectSettings);
-                if (isV2()) {
-                  assert.isTrue(ctx.contextV2 !== undefined);
-                }
-              }
-              return ok("");
-            }
-          }
-          hooks(MyClass, {
-            other: [ProjectSettingsLoaderMW, ContextInjectorMW],
-          });
-          const my = new MyClass();
-          const res = await my.other(inputs);
-          assert.isTrue(res.isOk() && res.value === "");
-        });
-
-        it(`case when missing activeEnvironment`, async () => {
-          class MyClass {
-            tools = new MockTools();
-            async other(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-              return ok("");
-            }
-          }
-          hooks(MyClass, {
-            other: [ProjectSettingsLoaderMW, ContextInjectorMW],
-          });
-          const my = new MyClass();
-          delete projectSettings.activeEnvironment;
-          const res = await my.other(inputs);
-          if (param.TEAMSFX_MULTI_ENV === "true") {
-            assert.isTrue(
-              res.isErr() &&
-                res.error.message.includes(
-                  `activeEnvironment is missing or not a string in ${ProjectSettingsFileName}`
-                )
-            );
-          } else {
-            assert.isTrue(res.isOk());
-          }
-          projectSettings.activeEnvironment = "dev";
-        });
-      });
-    }
-  });
-
-  describe("ProjectSettingsWriterMW", () => {
-    const sandbox = sinon.createSandbox();
-    afterEach(function () {
-      sandbox.restore();
-    });
-    it("ignore write", async () => {
-      const spy = sandbox.spy(fs, "writeFile");
-      class MyClass {
-        tools?: any = new MockTools();
-        async myMethod(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ProjectSettingsWriterMW],
-      });
-      const my = new MyClass();
-      const inputs1: Inputs = { platform: Platform.VSCode };
-      await my.myMethod(inputs1);
-      const inputs2: Inputs = {
-        platform: Platform.CLI_HELP,
-        projectPath: path.join(os.tmpdir(), randomAppName()),
-      };
-      await my.myMethod(inputs2);
-      const inputs3: Inputs = {
-        platform: Platform.VSCode,
-        projectPath: path.join(os.tmpdir(), randomAppName()),
-        ignoreConfigPersist: true,
-      };
-      await my.myMethod(inputs3);
-      const inputs4: Inputs = {
-        platform: Platform.VSCode,
-        projectPath: path.join(os.tmpdir(), randomAppName()),
-      };
-      await my.myMethod(inputs4);
-      assert(spy.callCount === 0);
-    });
-
-    it("write success for API V1 and API V2", async () => {
-      const appName = randomAppName();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), appName);
-      const tools = new MockTools();
-      const mockProjectSettings = MockProjectSettings(appName);
-      const fileMap = new Map<string, any>();
-      sandbox.stub<any, any>(fs, "writeFile").callsFake(async (file: string, data: any) => {
-        fileMap.set(file, data);
-      });
-      sandbox.stub(fs, "pathExists").resolves(true);
-      const confFolderPath = path.resolve(inputs.projectPath, `.${ConfigFolderName}`);
-      const settingsFileV1 = path.resolve(confFolderPath, "settings.json");
-      const settingsFileV2 = path.resolve(
-        confFolderPath,
-        InputConfigsFolderName,
-        ProjectSettingsFileName
-      );
-      class MyClass {
-        tools = tools;
-        async myMethod(inputs: Inputs, ctx?: CoreHookContext): Promise<Result<any, FxError>> {
-          if (ctx) ctx.projectSettings = mockProjectSettings;
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        myMethod: [ContextInjectorMW, ProjectSettingsWriterMW],
-      });
-      const my = new MyClass();
-      let mockedEnvRestore = mockedEnv({ TEAMSFX_MULTI_ENV: "false" });
-      {
-        await my.myMethod(inputs);
-        const content: string = fileMap.get(settingsFileV1);
-        const settingsInFile = JSON.parse(content);
-        assert.deepEqual(mockProjectSettings, settingsInFile);
-      }
-      mockedEnvRestore();
-      mockedEnvRestore = mockedEnv({ TEAMSFX_MULTI_ENV: "true" });
-      {
-        await my.myMethod(inputs);
-        const content: string = fileMap.get(settingsFileV2);
-        const settingsInFile = JSON.parse(content);
-        assert.deepEqual(mockProjectSettings, settingsInFile);
-      }
-      mockedEnvRestore();
-    });
-  });
-
-  describe("ProjectSettingsLoaderMW, ProjectSettingsWriterMW for user data encryption", () => {
-    const sandbox = sinon.createSandbox();
-
-    afterEach(function () {
-      sandbox.restore();
-    });
-
-    it("successfully encrypt userdata and load it", async () => {
-      const appName = randomAppName();
-      const inputs: Inputs = { platform: Platform.VSCode };
-      inputs.projectPath = path.join(os.tmpdir(), appName);
-      const tools = new MockTools();
-      const solutionContext = await newSolutionContext(tools, inputs);
-      const configMap = new ConfigMap();
-      const pluginName = "fx-resource-aad-app-for-teams";
-      const secretName = "clientSecret";
-      const secretText = "test";
-      configMap.set(secretName, secretText);
-      solutionContext.envInfo.profile.set("solution", new ConfigMap());
-      solutionContext.envInfo.profile.set(pluginName, configMap);
-      const oldProjectId = solutionContext.projectSettings!.projectId;
-      solutionContext.projectSettings = MockProjectSettings(appName);
-      solutionContext.projectSettings!.projectId = oldProjectId;
-      const fileMap = new Map<string, any>();
-      sandbox.stub<any, any>(fs, "writeFile").callsFake(async (file: string, data: any) => {
-        fileMap.set(file, data);
-      });
-      sandbox.stub(fs, "pathExists").resolves(true);
-
-      const envName = environmentManager.getDefaultEnvName();
-      const confFolderPath = path.resolve(inputs.projectPath, `.${ConfigFolderName}`);
-      const userdataFile = path.resolve(confFolderPath, `${envName}.userdata`);
-      const settingsFile = path.resolve(confFolderPath, "settings.json");
-      const envJsonFile = path.resolve(confFolderPath, `env.${envName}.json`);
-
-      class MyClass {
-        tools = tools;
-
-        async WriteConfigTrigger(
-          inputs: Inputs,
-          ctx?: CoreHookContext
-        ): Promise<Result<any, FxError>> {
-          ctx!.solutionContext = solutionContext;
-          return ok("");
-        }
-
-        async ReadConfigTrigger(
-          inputs: Inputs,
-          ctx?: CoreHookContext
-        ): Promise<Result<any, FxError>> {
-          assert.isTrue(ctx !== undefined);
-          assert.isTrue(ctx!.solutionContext !== undefined);
-          const solutionContext = ctx!.solutionContext!;
-          assert.isTrue(solutionContext.projectSettings !== undefined);
-          assert.isTrue(solutionContext.projectSettings!.appName === appName);
-          assert.isTrue(solutionContext.envInfo.profile.get(pluginName) !== undefined);
-          const value = solutionContext.envInfo.profile.get(pluginName)!.get(secretName);
-          assert.isTrue(value === secretText);
-          return ok("");
-        }
-      }
-
-      hooks(MyClass, {
-        WriteConfigTrigger: [ContextInjectorMW, ProjectSettingsWriterMW, EnvInfoWriterMW()],
-        ReadConfigTrigger: [ProjectSettingsLoaderMW, EnvInfoLoaderMW(true), ContextInjectorMW],
-      });
-      const my = new MyClass();
-      await my.WriteConfigTrigger(inputs);
-      const content = fileMap.get(userdataFile);
-      const userdata = dotenv.parse(content);
-      const secretValue = userdata[`${pluginName}.${secretName}`];
-      assert.isTrue(secretValue !== undefined);
-      assert.isTrue(secretValue.startsWith("crypto_"));
-
-      sandbox.stub<any, any>(fs, "readJson").callsFake(async (file: string) => {
-        if (settingsFile === file) return JSON.parse(fileMap.get(settingsFile));
-        if (envJsonFile === file) return JSON.parse(fileMap.get(envJsonFile));
-        return {};
-      });
-      sandbox.stub<any, any>(fs, "readFile").callsFake(async (file: string) => {
-        if (userdataFile === file) return content;
-        return {};
-      });
-      await my.ReadConfigTrigger(inputs);
-    });
-  });
-
-  describe("QuestionModelMW", () => {
-    const sandbox = sinon.createSandbox();
-    afterEach(function () {
-      sandbox.restore();
-    });
-
-    it("successful happy path", async () => {
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const tools = new MockTools();
-      const MockContextLoaderMW: Middleware = async (ctx: CoreHookContext, next: NextFunction) => {
-        ctx.solutionContext = await newSolutionContext(tools, inputs);
-        await next();
-      };
-
-      const ui = tools.ui;
-      const questionName = "mockquestion";
-      let questionValue = randomAppName();
-      sandbox.stub(ui, "inputText").callsFake(async (config: InputTextConfig) => {
-        return ok({ type: "success", result: questionValue });
-      });
-
-      class MockCore {
-        version = "1";
-        tools = tools;
-
-        async createProject(inputs: Inputs): Promise<Result<string, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async provisionResources(inputs: Inputs): Promise<Result<any, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async deployArtifacts(inputs: Inputs): Promise<Result<any, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async localDebug(inputs: Inputs): Promise<Result<any, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async publishApplication(inputs: Inputs): Promise<Result<any, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async executeUserTask(func: Func, inputs: Inputs): Promise<Result<unknown, FxError>> {
-          assert.isTrue(inputs[questionName] === questionValue);
-          return ok("");
-        }
-
-        async _getQuestionsForCreateProject(
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          const node = new QTreeNode({
-            type: "text",
-            name: questionName,
-            title: "",
-          });
-          return ok(node);
-        }
-
-        async _getQuestions(
-          ctx: SolutionContext,
-          solution: Solution,
-          stage: Stage,
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          const node = new QTreeNode({
-            type: "text",
-            password: true,
-            name: questionName,
-            title: "",
-          });
-          return ok(node);
-        }
-
-        async _getQuestionsForUserTask(
-          ctx: SolutionContext,
-          solution: Solution,
-          func: FunctionRouter,
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          const node = new QTreeNode({
-            type: "text",
-            name: questionName,
-            title: "",
-          });
-          return ok(node);
-        }
-      }
-
-      hooks(MockCore, {
-        createProject: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        provisionResources: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        deployArtifacts: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        localDebug: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        publishApplication: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        executeUserTask: [SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-      });
-      const my = new MockCore();
-
-      const res = await my.createProject(inputs);
-      assert.isTrue(res.isOk() && res.value === "");
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "provisionResources";
-      await my.provisionResources(inputs);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "deployArtifacts";
-      await my.deployArtifacts(inputs);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "localDebug";
-      await my.localDebug(inputs);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "publishApplication";
-      await my.publishApplication(inputs);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "executeUserTask";
-      const func: Func = { method: "test", namespace: "" };
-      await my.executeUserTask(func, inputs);
-    });
-
-    it("get question or traverse question tree error", async () => {
-      const inputs: Inputs = { platform: Platform.VSCode };
-      const tools = new MockTools();
-      const MockContextLoaderMW: Middleware = async (ctx: CoreHookContext, next: NextFunction) => {
-        ctx.solutionContext = await newSolutionContext(tools, inputs);
-        await next();
-      };
-
-      const ui = tools.ui;
-      const questionName = "mockquestion";
-      let questionValue = randomAppName();
-      sandbox.stub(ui, "inputText").callsFake(async (config: InputTextConfig) => {
-        return ok({ type: "success", result: questionValue });
-      });
-
-      class MockCore {
-        version = "1";
-        tools = tools;
-
-        async createProject(inputs: Inputs): Promise<Result<string, FxError>> {
-          return ok("");
-        }
-
-        async provisionResources(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-
-        async deployArtifacts(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-
-        async localDebug(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-
-        async publishApplication(inputs: Inputs): Promise<Result<any, FxError>> {
-          return ok("");
-        }
-
-        async executeUserTask(func: Func, inputs: Inputs): Promise<Result<unknown, FxError>> {
-          return ok("");
-        }
-
-        async _getQuestionsForCreateProject(
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          return err(InvalidInputError("mock"));
-        }
-
-        async _getQuestions(
-          ctx: SolutionContext,
-          solution: Solution,
-          stage: Stage,
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          return err(InvalidInputError("mock"));
-        }
-
-        async _getQuestionsForUserTask(
-          ctx: SolutionContext,
-          solution: Solution,
-          func: FunctionRouter,
-          inputs: Inputs
-        ): Promise<Result<QTreeNode | undefined, FxError>> {
-          const node = new QTreeNode({
-            type: "singleSelect",
-            name: questionName,
-            title: "",
-            staticOptions: [],
-          });
-          return ok(node);
-        }
-      }
-
-      hooks(MockCore, {
-        createProject: [ErrorHandlerMW, SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        provisionResources: [
-          ErrorHandlerMW,
-          SolutionLoaderMW(),
-          MockContextLoaderMW,
-          QuestionModelMW,
-        ],
-        deployArtifacts: [ErrorHandlerMW, SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        localDebug: [ErrorHandlerMW, SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-        publishApplication: [
-          ErrorHandlerMW,
-          SolutionLoaderMW(),
-          MockContextLoaderMW,
-          QuestionModelMW,
-        ],
-        executeUserTask: [ErrorHandlerMW, SolutionLoaderMW(), MockContextLoaderMW, QuestionModelMW],
-      });
-      const my = new MockCore();
-
-      let res = await my.createProject(inputs);
-      assert(res.isErr() && res.error.name === InvalidInputError("").name);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "provisionResources";
-      res = await my.provisionResources(inputs);
-      assert(res.isErr() && res.error.name === InvalidInputError("").name);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "deployArtifacts";
-      res = await my.deployArtifacts(inputs);
-      assert(res.isErr() && res.error.name === InvalidInputError("").name);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "localDebug";
-      res = await my.localDebug(inputs);
-      assert(res.isErr() && res.error.name === InvalidInputError("").name);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "publishApplication";
-      res = await my.publishApplication(inputs);
-      assert(res.isErr() && res.error.name === InvalidInputError("").name);
-
-      delete inputs[questionName];
-      questionValue = randomAppName() + "executeUserTask";
-      const func: Func = { method: "test", namespace: "" };
-      const res2 = await my.executeUserTask(func, inputs);
-      assert(res2.isErr() && res2.error.name === "EmptySelectOption");
-    });
   });
 
   describe("ProjectUpgraderMW", () => {
